Added tests

+++ loan.calculator/src/main/java/interview/loan/calculator/repository/LoanRepository.java
@@ -5,6 +5,6 @@ import org.springframework.data.jpa.repository.JpaRepository;
 import org.springframework.stereotype.Repository;
 
 @Repository
-public interface LoanRepository extends JpaRepository<PaymentInfoEntity, Integer> {
+public interface LoanRepository extends JpaRepository<PaymentInfoEntity, Long> {
 
 }
